Added Item CRUD

+++ BackEnd/src/pos/dto/PurchaseOrderDTO.java
@@ -7,7 +7,7 @@ public class PurchaseOrderDTO {
     String oid;
     String code;
     int qty;
-    BigDecimal unitPrice;
+    double unitPrice;
 
     public String getOid() {
         return oid;
@@ -21,7 +21,7 @@ public class PurchaseOrderDTO {
         return qty;
     }
 
-    public BigDecimal getUnitPrice() {
+    public double getUnitPrice() {
         return unitPrice;
     }
 
@@ -37,14 +37,14 @@ public class PurchaseOrderDTO {
         this.qty = qty;
     }
 
-    public void setUnitPrice(BigDecimal unitPrice) {
+    public void setUnitPrice(double unitPrice) {
         this.unitPrice = unitPrice;
     }
 
     public PurchaseOrderDTO() {
     }
 
-    public PurchaseOrderDTO(String oid, String code, int qty, BigDecimal unitPrice) {
+    public PurchaseOrderDTO(String oid, String code, int qty, double unitPrice) {
         this.oid = oid;
         this.code = code;
         this.qty = qty;
